feat: add version and ignore validation err flags
Adds a version flag that will return the version of schemacheck. The
version is set automatically by GoReleaser upon build through a default
ldflag. Adds a ignore-val-err flag that will warn and exit with exit
code 0 instead of 1. This exit only affects when the document exists but
doesn't return valid according to the schema. May be useful for when you want a
soft failure for historical systems.